webapp/requirements.txt
aspy.yaml==1.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
cfgv==2.0.1
chardet==3.0.4
click==6.7
coverage==5.0.3
cryptography==2.8
fabric==2.5.0
Flask==1.1.1
Flask-Cache==0.13.1
Flask-Humanize==0.3.0
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-datastore==1.10.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.17.0
humanize==0.5.1
identify==1.4.11
idna==2.5
importlib-metadata==1.5.0
invoke==1.4.1
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
nodeenv==1.3.4
oauth2client==4.1.3
oauthlib==3.0.1
paramiko==2.7.1
pre-commit==2.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.3
PyYAML==5.2
requests==2.18.1
rsa==4.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
stripe==2.42.0
toml==0.10.0
uritemplate==3.0.1
urllib3==1.21.1
virtualenv==16.7.9
Werkzeug==0.15.5
WTForms==2.2.1
zipp==2.1.0
